Moved node implementation into dedicated source file
